requirements.txt
streamlit>=1.28.0,<2.0.0
boto3>=1.34.0,<2.0.0
pandas>=2.0.0,<3.0.0
plotly>=5.17.0,<6.0.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
setuptools>=65.0.0
psutil>=5.9.0,<6.0.0
psutil>=5.9.0,<6.0.0